requirements.txt
onnx
numpy
scikit-learn
matplotlib
zstandard
tqdm
pycapnp